requirements.txt
streamlit
xgboost
scikit-learn
pandas
numpy
scapy
matplotlib
joblib
fpdf2